[Java] Implement PR feedback.
